move core auth out of identity file
