+++ compute_space/src/compute_space/web/auth/pages.py
@@ -17,16 +17,16 @@ from quart import url_for
 from quart.typing import ResponseReturnValue
 
 from compute_space.config import get_config
-from compute_space.core.auth.cookies import COOKIE_ACCESS
-from compute_space.core.auth.cookies import COOKIE_REFRESH
-from compute_space.core.auth.cookies import clear_auth_cookies
-from compute_space.core.auth.cookies import set_auth_cookies
-from compute_space.core.auth.identity import get_current_user_from_request
+from compute_space.core.auth.auth import get_current_user_from_request
 from compute_space.core.auth.tokens import REFRESH_TOKEN_EXPIRY
 from compute_space.core.auth.tokens import create_access_token
 from compute_space.core.default_apps import deploy_default_apps
 from compute_space.core.logging import logger
 from compute_space.db import get_db
+from compute_space.web.auth.cookies import COOKIE_ACCESS
+from compute_space.web.auth.cookies import COOKIE_REFRESH
+from compute_space.web.auth.cookies import clear_auth_cookies
+from compute_space.web.auth.cookies import set_auth_cookies
 from compute_space.web.auth.middleware import _try_refresh  # noqa: F401 — re-exported
 from compute_space.web.auth.middleware import login_required  # noqa: F401 — re-exported
